split out to custom hook and add redirect check
